+++ Src/Core/VulkanRenderer.h
@@ -107,7 +107,7 @@ namespace VkRender {
         virtual void viewChanged();
 
         /** @brief (Virtual) Called after the mouse cursor moved and before internal events (like camera m_Rotation) is firstUpdate */
-        virtual void mouseMoved(double x, double y, bool &handled);
+        virtual void mouseMoved(float x, float y, bool &handled);
 
         /** @brief (Virtual) Called when the window has been resized, can be used by the sample application to recreate resources */
         virtual void windowResized();
@@ -230,7 +230,7 @@ namespace VkRender {
         void setWindowSize(uint32_t width, uint32_t height);
 
         /** @brief Default function to handle cursor m_Position input, calls the override function mouseMoved(...) **/
-        void handleMouseMove(int32_t x, int32_t y);
+        void handleMouseMove(float x, float y);
 
         [[nodiscard]] VkPhysicalDevice pickPhysicalDevice(std::vector<VkPhysicalDevice> devices) const;
